+++ CHANGES.rst
@@ -10,6 +10,7 @@ Added
   phase, they get redirected back to that page afterwards.
   :issue:`168` :pr:`151`
 - flask-babel and pytz are now part of the `front` extras
+- Bump to fomantic-ui 2.9.3 :pr:`152`
 
 [0.0.33] - 2023-08-26
 =====================