+++ CHANGES.rst
@@ -17,6 +17,7 @@ Added
   :issue:`168` :pr:`151`
 - flask-babel and pytz are now part of the `front` extras
 - Bump to fomantic-ui 2.9.3 :pr:`152`
+- Bump to htmx 1.9.6 :pr:`154`
 
 [0.0.33] - 2023-08-26
 =====================