+++ CHANGES.rst
@@ -15,6 +15,11 @@ Changed
 - Model attributes cardinality is closer to SCIM model. :pr:`155`
 - Bump to htmx 1.9.9 :pr:`159`
 
+Fixed
+*****
+
+- Disable HTMX boosting during the OIDC dance. :pr:`160`
+
 [0.0.34] - 2023-10-02
 =====================