+++ CHANGES.rst
@@ -3,11 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+[0.0.35] - 2023-11-25
+=====================
+
 Added
 *****
 
 - Refresh token grant supports other client authentication methods. :pr:`157`
-- Implement a SQLAlchemy backend. :pr:`158`
+- Implement a SQLAlchemy backend. :issue:`30` :pr:`158`
 
 Changed
 *******